requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-WTF==1.2.1
Flask-Limiter==3.5.0
SQLAlchemy==2.0.23
Werkzeug==3.0.1
WTForms==3.1.1
email-validator==2.1.0.post1
python-dotenv==1.0.0
Pillow==10.1.0
stripe==7.6.0
requests==2.31.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
